tools/c7n_dispatcher/requirements.txt
boto3
botocore
c7n
ElasticMock
flake8
jsonschema
mock
nose
pytest-cov
pytest-xdist
pytest
PyYAML
six
tox